SCMultiSliderView: nThumbSize mThumbSizeY changed to float
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@4425 a380766d-ff14-0410-b294-a243070f3f08

+++ headers/app/SCView.h
@@ -565,7 +565,7 @@ protected:
 	double valueFromPoint(SCPoint where);
 	void setValueFromPoint(SCPoint point);
 	SCRect calcThumbRect(int xIn, double valIn, double step);
-	int mThumbSize, mThumbSizeY; // size of the rect
+	float mThumbSize, mThumbSizeY; // size of the rect
 	int mTabSize, mVisibleSize; // size of the table
 	SCColor mFillColor;
 	SCColor mStrokeColor;
